refactor(SearchInput): aligning files names with component name, adding input props
