+++ packages/react-components/src/components/SearchInput/SearchInput.mdx
@@ -1,10 +1,10 @@
 import { Meta, Title, Canvas, ArgTypes } from '@storybook/blocks';
 
-import * as Stories from './Search.stories';
+import * as Stories from './SearchInput.stories';
 
 <Meta of={Stories} />
 
-<Title>Search</Title>
+<Title>SearchInput</Title>
 
 [Intro](#Intro) | [Component API](#ComponentAPI) | [Content Spec](#ContentSpec)
