+++ packages/react-components/src/components/SearchInput/SearchInput.module.scss
